requirements.txt
numpy<2
pandas>=2.1,<2.3
xarray>=2024.1.0,<2025.0.0
netCDF4>=1.6.5
h5netcdf>=1.3.0
PyYAML>=6.0
torch>=2.2
matplotlib>=3.8
